first pass at form validation
Using Jquery validation for form validation. It is working but there is
a issue with the sign in from the top nag bar for some reason. Other
than that all initial form validation is done. Still need to work on
not allowing users to double add semesters, courses or assignments and
to clean up input